Intialization
So, we have a project to submit for a class called business intelligence and analytics. For this class, I thought let me do some data engineering.

Hence this project. This is basically pulling may API, Docs and Excels from many places and then saving them in a structred way. So, i can use it for analytics. Which can be converted into a dashboard.