CEA data scraping
Created a CEA data scraping script, which saves excel files inside the CEA reports page into two folders. Which are daily and monthly data. As the website is a static website with <a> tag reference. So, i just used bs4 for scraping and save the files in the name of titles.